requirements.txt
streamlit
pandas
numpy
matplotlib
tensorflow=2.15.0
pillow